backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
Werkzeug==3.1.2
Authlib==1.3.2
